Implement generic routing. Bump version to 0.1.1.
You can now specify any number of routes on the command line,
with the option -r. The target for a route can either be a file
or directory, a Nix request handler or static response code.
nixrbd will always select the most specific route for a request. It
has a default route for / which maps to the response code 404, not found.
Hence, any requests that don't match the routes you have setup will
get a 404 response.

The options -d and -m has been removed in favour for the route option.